requirements.txt
playwright>=1.40.0
pyyaml>=6.0
sqlalchemy>=2.0
streamlit>=1.30.0
pandas>=2.1
plotly>=5.18
python-dotenv>=1.0
